// cypress/support/pageObjects/dataEntryForm/enums/dataEntryFormEnum.ts
export enum CountriesEnum {
    USA = "USA",
    CANADA = "Canada",
    UNITED_KINGDOM = "United Kingdom",
    GERMANY = "Germany",
    FRANCE = "France",
    AUSTRALIA = "Australia",
    JAPAN = "Japan",
    BRAZIL = "Brazil",
    CHINA = "China",
    INDIA = "India",
}

export enum ColorsEnum {
    RED = "Red",
    BLUE = "Blue",
    GREEN = "Green",
    YELLOW = "Yellow",
    WHITE = "White",
}

export enum SortedListEnum {
    CAT = "Cat",
    CHEETAH = "Cheetah",
    DEER = "Deer",
    DOG = "Dog",
    ELEPHANT = "Elephant",
    FOX = "Fox",
    GIRAFFE = "Giraffe",
    LION = "Lion",
    RABBIT = "Rabbit",
    ZEBRA = "Zebra"
}